refactor(discord): compress comments and pure-move panel helpers
Reduce namespace-cap pressure without changing signatures or control flow: compress streaming tick comments and move cohesive status derivation, cache invalidation facades, and completion cleanup helpers into child modules.\n\nCo-Authored-By: Claude Fable 5 <noreply@anthropic.com>
